Extract hardware_helper.rs
